🔄✨ Refactor state management: Separate useImageExportStore and useImageNodesStore
- Separated state management functions into two distinct files: useImageExportStore.ts and useImageNodesStore.ts.
- Updated all relevant parts of the Figma plugin code to reflect these changes.
