+++ plugin/src/app/store/useImageExportStore.ts
@@ -1,10 +1,8 @@
 import create from "zustand";
 
 // ** import types
-import { NodeData } from "@/types/node";
-import { ImageExportState } from "@/types/state";
+import { ImageExportState } from "@/types/export-settings";
 import { ExportOption, ExportScaleOption, CaseOption } from "@/types/enums";
-
 
 export const useImageExportStore = create<ImageExportState>((set) => ({
   exportOption: ExportOption.JPG,
@@ -13,21 +11,6 @@ export const useImageExportStore = create<ImageExportState>((set) => ({
   exportScaleOption: ExportScaleOption.ONE_X,
   setExportScaleOption: (option: ExportScaleOption) =>
     set({ exportScaleOption: option }),
-
-  allNodes: [],
-  setAllNodes: (nodes: NodeData[]) => set({ allNodes: nodes }),
-
-  selectedNodeIds: [],
-  setSelectedNodeIds: (ids: string[] | ((prev: string[]) => string[])) =>
-    set((state) => ({
-      selectedNodeIds: typeof ids === 'function' ? ids(state.selectedNodeIds) : ids,
-    })),
-
-  allNodesCount: 0,
-  setAllNodesCount: (count: number) => set({ allNodesCount: count }),
-
-  selectedNodesCount: 0,
-  setSelectedNodesCount: (count: number) => set({ selectedNodesCount: count }),
 
   caseOption: CaseOption.CAMEL_CASE,
   setCaseOption: (option: CaseOption) => set({ caseOption: option }),
