+++ plugin/src/app/types/export-settings.ts
@@ -1,3 +1,5 @@
+import { CaseOption, ExportOption, ExportScaleOption } from "./enums";
+
 export interface ExportSettings {
     format: string;
     constraint?: { type: "SCALE", value: number };
@@ -14,4 +16,15 @@ export interface ExportSettings {
   export interface ExportSettingsSVG extends ExportSettings {
     format: "SVG";
   }
-  
+
+
+export interface ImageExportState {
+  exportOption: ExportOption;
+  setExportOption: (option: ExportOption) => void;
+
+  exportScaleOption: ExportScaleOption;
+  setExportScaleOption: (option: ExportScaleOption) => void;
+
+  caseOption: CaseOption;
+  setCaseOption: (option: CaseOption) => void;
+}
